Table: move TableNG hooks into a hooks/ folder
Same treatment as utils/: hooks.ts had reached 999 lines spanning the row
pipeline, height measurement, column widths, resize interaction and the footer.
Split into four modules under hooks/ with a barrel index.ts, so the existing
`./hooks` and `../hooks` import sites resolve unchanged:

  rows    filtering, sorting, pagination, frame compilation, nested rows
  height  the typography context, header height and row heights
  width   column widths, user-driven resize, scrollbar width
  footer  reducer entries for the summary row

useTypographyCtx sits in height.ts because text measurement is what it exists
for, even though the width hooks also take a TypographyCtx as an argument.

render-hooks.tsx moves in whole as hooks/render.tsx rather than being split:
these are data-grid hooks like the rest, and leaving them in a sibling file while
everything else moved into hooks/ read arbitrarily. Only its relative import
depths change.

That move needed one fix to stay acyclic. hooks/render.tsx imports
components/SummaryCell, which imported the hooks barrel, so exporting render from
index.ts would have made hooks/index.ts circular. SummaryCell now imports
useReducerEntries from ../hooks/footer directly.

The hooks also now import the table utilities from their leaf modules
(../utils/rows, ../utils/height, and so on) rather than through the utils barrel.

hooks.test.ts and render.test.tsx are left untouched here: the existing suites
passing against the new module graph is what shows the move is behaviour
preserving. hooks.test.ts is split in the follow-up commit.
