+++ packages/grafana-ui/src/components/Table/TableNG/hooks/index.ts
@@ -1,0 +1,7 @@
+// Barrel for the TableNG data-grid hooks. Re-exports only: modules in this folder import their
+// siblings by direct path so the barrel never pulls the whole folder into a single leaf.
+export * from './footer';
+export * from './height';
+export * from './render';
+export * from './rows';
+export * from './width';
